Use wlr_output_swapchain
References: https://gitlab.freedesktop.org/wlroots/wlroots/-/merge_requests/3984
